\name{hot-package}
\alias{hot-package}
\alias{hot}
\docType{package}
\title{Hotelling T2 }
\description{Computes Hotelling \eqn{T^2} statistics for microarray}

\details{
\tabular{ll}{
Package: \tab hot\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2005-12-05\cr
License: \tab GPL\cr
}
Computes Hotelling  statistics for microarray
}
\author{Gilles Guillot

  Maintainer: Gilles Guillot <guillot@inapg.fr>
 
  
}
\references{Discrimination and scoring  using small sets of genes for 
    two-sample microarray data,
    G. Guillot, M. Olsson, M. Bensson, M. Rudemo, submitted. }

\keyword{ }




